Script file support added & Major UI changes
> ScriptHandler.kt created to handle all tasks related to scripts.

> Script files (.txt for simplicity)  can now be selected and executed through the updated UI.

> 'Exec' -> 'Execute as Script' button

> All UI components below the TouchPad is now in a scrollable box
