+++ app/src/main/java/me/arianb/usb_hid_client/input_views/ManualInput.kt
@@ -3,7 +3,9 @@ package me.arianb.usb_hid_client.input_views
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.wrapContentSize
 import androidx.compose.material3.Button
 import androidx.compose.material3.Text
@@ -17,14 +19,16 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
 import me.arianb.usb_hid_client.MainViewModel
 import me.arianb.usb_hid_client.R
 import me.arianb.usb_hid_client.hid_utils.KeyCodeTranslation
 import me.arianb.usb_hid_client.settings.SettingsViewModel
 import me.arianb.usb_hid_client.ui.theme.PaddingLarge
+import me.arianb.usb_hid_client.ui.theme.PaddingNormal
+import me.arianb.usb_hid_client.ui.theme.PaddingSmall
 import timber.log.Timber
-import java.util.Locale.getDefault
 
 @Composable
 fun ManualInput(
@@ -35,27 +39,31 @@ fun ManualInput(
     val preferencesState by settingsViewModel.userPreferencesFlow.collectAsState()
     val shouldClearManualInputOnSend = preferencesState.clearManualInput
 
-    Row(
-        modifier = Modifier.fillMaxWidth(),
-        horizontalArrangement = Arrangement.spacedBy(PaddingLarge),
-        verticalAlignment = Alignment.CenterVertically
+    Column(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(0.dp, PaddingLarge,0.dp,0.dp),
+        verticalArrangement = Arrangement.spacedBy(PaddingSmall),
+        horizontalAlignment = Alignment.CenterHorizontally
     ) {
         TextField(
             modifier = Modifier
-                .fillMaxWidth()
-                .weight(1f), // For some reason, this makes the button not be squished at the end of the Row
+                .fillMaxWidth(),
             value = manualInputString,
             label = { Text(stringResource(R.string.manual_input)) },
             onValueChange = { manualInputString = it }
         )
 
-        Column (
-            modifier = Modifier.wrapContentSize(),
-            verticalArrangement = Arrangement.spacedBy(PaddingLarge),
-            horizontalAlignment = Alignment.CenterHorizontally
+        Row (
+            modifier = Modifier
+                .fillMaxWidth(),
+            horizontalArrangement = Arrangement.End,
+            verticalAlignment = Alignment.CenterVertically
         ) {
             Button( // pass to sendString()
-                modifier = Modifier.wrapContentSize(),
+                modifier = Modifier
+                    .wrapContentSize()
+                    .padding(0.dp,0.dp,PaddingNormal,0.dp),
                 onClick = onClick@{
                     // If empty, don't do anything
                     if (manualInputString.isEmpty()) {
@@ -222,103 +230,4 @@ fun sendInput(stringToSend: String, mainViewModel: MainViewModel) {
 
 }
 
-fun scriptExecutor(script: String, mainViewModel: MainViewModel) {
-    //---------- Contributed by saaiqSAS ----------
-    // This method allows the use of scripting. The entire script should be passed to this method and it will be executed line by line
 
-    val lines = (script).split("\n")
-
-
-    Timber.d("num of lines: " + lines.size) //test
-
-    for (line in lines) {
-        val command: String
-        var key = ""
-        var para = ""
-        val firstSpace = line.indexOf(" ")
-
-        Timber.d("line: " + line) //test
-
-        if (firstSpace == -1) {
-            command = line
-        } else {
-            command = line.take(firstSpace)
-            para = line.substring(firstSpace + 1)
-        }
-
-        Timber.d("command: $command") //test
-        Timber.d("para: $para") //test
-
-        when (command.uppercase(getDefault())) {
-            //COMMANDS - only commands takes a parameter
-            "//","REM","NAME","DESC","AUTHOR"   -> {} //do nothing
-            "SEND", "STRING"                    -> sendInput(para, mainViewModel)
-            "SENDLN", "STRINGLN"                -> sendInput(para+"\n", mainViewModel)
-            "SLEEP","DELAY"                     -> Thread.sleep(para.toLong())
-
-            //MODIFIER KEYS
-            "L_CTRL", "L_CONTROL", "CTRL"       -> key = "left-ctrl"
-            "L_ALT", "ALT"                      -> key = "left-alt"
-            "L_SHIFT", "SHIFT"                  -> key = "left-shift"
-            "L_META", "L_WIN", "WIN"            -> key = "left-meta"
-            "R_CTRL", "R_CONTROL"               -> key = "right-ctrl"
-            "R_ALT"                             -> key = "right-alt"
-            "R_SHIFT"                           -> key = "right-shift"
-            "R_META", "R_WIN"                   -> key = "right-meta"
-
-            //SPECIAL KEYS
-            "UP"                                -> key = "up"
-            "DOWN"                              -> key = "down"
-            "LEFT"                              -> key = "left"
-            "RIGHT"                             -> key = "right"
-            "ESCAPE", "ESC"                     -> key = "escape"
-            "TAB"                               -> key = "tab"
-            "BACKSPACE","BACK"                  -> key = "backspace"
-            "DELETE","DEL"                      -> key = "delete"
-            "PRINT"                             -> key = "print"
-            "SPACE"                             -> key = " "
-            "ENTER"                             -> key = "\n"
-            "SCROLL_LOCK"                       -> key = "scroll-lock"
-            "NUM_LOCK"                          -> key = "num-lock"
-            "PAUSE"                             -> key = "pause"
-            "INSERT"                            -> key = "insert"
-            "HOME"                              -> key = "home"
-            "END"                               -> key = "end"
-            "PAGE_UP", "PG_UP"                  -> key = "page-up"
-            "PAGE_DOWN", "PG_DOWN"              -> key = "page-down"
-            "NEXT"                              -> key = "next"
-            "PREVIOUS", "PREV"                  -> key = "previous"
-            "PLAY_PAUSE", "PLAY","PP"           -> key = "play-pause"
-            "VOLUME_UP", "VOL_UP"               -> key = "volume-up"
-            "VOLUME_DOWN", "VOL_DOWN"           -> key = "volume-down"
-            "F1"                                -> key = "f1"
-            "F2"                                -> key = "f2"
-            "F3"                                -> key = "f3"
-            "F4"                                -> key = "f4"
-            "F5"                                -> key = "f5"
-            "F6"                                -> key = "f6"
-            "F7"                                -> key = "f7"
-            "F8"                                -> key = "f8"
-            "F9"                                -> key = "f9"
-            "F10"                               -> key = "f10"
-            "F11"                               -> key = "f11"
-            "F12"                               -> key = "f12"
-
-
-        }
-        if (!key.isEmpty()) {
-            val scanCodes = KeyCodeTranslation.keyCharToScanCodes(key)
-
-            if (scanCodes == null) {
-                val error = "key: '$key' is not supported."
-                Timber.e(error)
-                return
-            }
-
-            if (scanCodes.second != 0x0.toByte()) {
-                mainViewModel.addStandardKey(scanCodes.first, scanCodes.second)
-            }
-        }
-    }
-
-}
