+++ app/src/main/res/values/strings.xml
@@ -7,7 +7,7 @@
     <string name="manual_input">Manual Input</string> <!-- Reused as preference category header -->
     <string name="touchpad_label">Touchpad</string> <!-- Reused as preference category header -->
     <string name="send">Send</string>
-    <string name="exec">Exec</string>
+    <string name="exec">Execute as Script</string>
 
     <!-- Settings Screen -->
     <string name="settings">Settings</string>
